Struct pkcs8::PrivateKeyInfo

pub struct PrivateKeyInfo<'a> {
    pub algorithm: AlgorithmIdentifier<'a>,
    pub private_key: &'a [u8],
    pub public_key: Option<&'a [u8]>,
}

PKCS#8 PrivateKeyInfo.

ASN.1 structure containing an AlgorithmIdentifier, private key data in an algorithm specific format, and optional attributes (ignored by this implementation).

Supports PKCS#8 v1 as described in RFC 5208 and PKCS#8 v2 as described in RFC 5958. PKCS#8 v2 keys include an additional public key field.

PKCS#8 v1 PrivateKeyInfo

Described in RFC 5208 Section 5:

PrivateKeyInfo ::= SEQUENCE {
        version                   Version,
        privateKeyAlgorithm       PrivateKeyAlgorithmIdentifier,
        privateKey                PrivateKey,
        attributes           [0]  IMPLICIT Attributes OPTIONAL }

Version ::= INTEGER

PrivateKeyAlgorithmIdentifier ::= AlgorithmIdentifier

PrivateKey ::= OCTET STRING

Attributes ::= SET OF Attribute

PKCS#8 v2 OneAsymmetricKey

PKCS#8 OneAsymmetricKey as described in RFC 5958 Section 2:

PrivateKeyInfo ::= OneAsymmetricKey

OneAsymmetricKey ::= SEQUENCE {
    version                   Version,
    privateKeyAlgorithm       PrivateKeyAlgorithmIdentifier,
    privateKey                PrivateKey,
    attributes            [0] Attributes OPTIONAL,
    ...,
    [[2: publicKey        [1] PublicKey OPTIONAL ]],
    ...
  }

Version ::= INTEGER { v1(0), v2(1) } (v1, ..., v2)

PrivateKeyAlgorithmIdentifier ::= AlgorithmIdentifier

PrivateKey ::= OCTET STRING

Attributes ::= SET OF Attribute

PublicKey ::= BIT STRING

Fields
algorithm: AlgorithmIdentifier<'a>

X.509 AlgorithmIdentifier for the private key type.

private_key: &'a [u8]

Private key data.

public_key: Option<&'a [u8]>

Public key data, optionally available if version is V2.

Implementations

impl<'a> PrivateKeyInfo<'a>

pub fn new(algorithm: AlgorithmIdentifier<'a>, private_key: &'a [u8]) -> Self

Create a new PKCS#8 PrivateKeyInfo message.

This is a helper method which initializes attributes and public_key to None, helpful if you aren’t using those.

pub fn version(&self) -> Version

Get the PKCS#8 Version for this structure.

Version::V1 if public_key is None, Version::V2 if Some.