Trait pkcs8::DecodePrivateKey

pub trait DecodePrivateKey: Sized {
    // Required method
    fn from_pkcs8_der(bytes: &[u8]) -> Result<Self>;

    // Provided methods
    fn from_pkcs8_encrypted_der(
        bytes: &[u8],
        password: impl AsRef<[u8]>
    ) -> Result<Self> { ... }
    fn from_pkcs8_pem(s: &str) -> Result<Self> { ... }
    fn from_pkcs8_encrypted_pem(
        s: &str,
        password: impl AsRef<[u8]>
    ) -> Result<Self> { ... }
    fn read_pkcs8_der_file(path: impl AsRef<Path>) -> Result<Self> { ... }
    fn read_pkcs8_pem_file(path: impl AsRef<Path>) -> Result<Self> { ... }
}

Parse a private key object from a PKCS#8 encoded document.

Required Methods

fn from_pkcs8_der(bytes: &[u8]) -> Result<Self>

Deserialize PKCS#8 private key from ASN.1 DER-encoded data (binary format).

Provided Methods

fn from_pkcs8_encrypted_der( bytes: &[u8], password: impl AsRef<[u8]> ) -> Result<Self>

Deserialize encrypted PKCS#8 private key from ASN.1 DER-encoded data (binary format) and attempt to decrypt it using the provided password.

fn from_pkcs8_pem(s: &str) -> Result<Self>

Deserialize PKCS#8-encoded private key from PEM.

Keys in this format begin with the following delimiter:

-----BEGIN PRIVATE KEY-----

fn from_pkcs8_encrypted_pem(s: &str, password: impl AsRef<[u8]>) -> Result<Self>

Deserialize encrypted PKCS#8-encoded private key from PEM and attempt to decrypt it using the provided password.

Keys in this format begin with the following delimiter:

-----BEGIN ENCRYPTED PRIVATE KEY-----

fn read_pkcs8_der_file(path: impl AsRef<Path>) -> Result<Self>

Load PKCS#8 private key from an ASN.1 DER-encoded file on the local filesystem (binary format).

fn read_pkcs8_pem_file(path: impl AsRef<Path>) -> Result<Self>

Load PKCS#8 private key from a PEM-encoded file on the local filesystem.

Object Safety

This trait is not object safe.